refactor: split repository into domain modules, fix credentials_file overlay bug, remove dead code
- Split 827-line repository.py into 5 domain modules under db/repository/
- Fix bug: credentials_file missing from DB overlay list (lost on restart)
- Remove dead duration_filter.py (unused, logic inlined in pipeline)
- Update 9 consumer imports and tests
- All 112 tests pass, ruff clean
