+++ .gitignore
@@ -15,3 +15,4 @@ __pycache__
 *.egg-info
 docs/superpowers/
 .opencode/
+graphify-out
