+++ src/ys2wl/api/routes/rules.py
@@ -2,7 +2,7 @@ import logging
 from typing import List
 from fastapi import APIRouter, HTTPException, Request
 from ys2wl.api.models import RoutingRuleCreate, RoutingRuleUpdate, RoutingRuleResponse
-from ys2wl.db import repository as repo
+from ys2wl.db.repository import pipeline as repo
 
 log = logging.getLogger("ys2wl.api.rules")
 router = APIRouter()
